requirements.txt
pandas==1.5.2
mesa==1.1.1
scikit-learn==1.1.3
networkx==2.8.8
numpy==1.23.5
scikit-surprise==1.1.3
